Split operators gather
This splits the cluster operators gather into two gathers, one for operator resources and other for unhealty operators.
